12 Jan 2002   S. A. Clough
I have developed a uniform block for the physical constants for all the modules:
  /CONSTS/ and have updated the values therein to current (1986) NIST values.
The values are set in the lblrtm.f module.

In so doing, I have had to change the /constn/ common block in lblatm.f, now /CNSTATM/

and the /CONSTL/ common block in the lbllow.f module (name unchanged).

This provides a uniform method for knowing the values for the physical constants used
everwhere (almost) in the program and for updating these values.
______________________________________________________________________________________

In lblatm.f, the write to TAPE7 (ipu = 1)has been changed to enable the sequential
writing of multiple profiles to this file, consistent with how TAPE6 is treated.  I
have also added to TAPE7 the profile number, the associated header record and the
accociated profile information.

______________________________________________________________________________________

I have changed all the alog and dlog refernces to ' 'log in all the modules where ' '
stands for blank.  This intended to address a problem with the double and single
options on some compilers.

18 Jan 2002 M. Shephard

Added a fix for when one reads in precalculated tangent KFILE (with imrg =22) so that
JPATHL=2, which then is set to the IPATHL in EMINT.

 c     This occurs when you are reading in precalculated KFILE
 c     with imrg options for merged (not sequential) output
 c     imrg=12 upwelling, imrg=22 tangent, imrg=32 downwelling
 C
       IF (IMRG.NE.2) THEN
          IF (IMRG.EQ.12) THEN
             JPATHL = 1
          else if (imrg.eq.22) then
             JPATHL = 2
>          ELSE
             JPATHL = 3
         ENDIF