MWS 04/02
Modified the format of the line rejection files REJ1 and REJ4 in lblrtm.f
They now are outputted and read in on layers with dv also beening stored and
read in.  This allow for consistent line coupling rejection from run to run for
the same case.

MWS 05/02

Fixed small emissivity/reflectivity (mostly interpolation) problems in
pointed out by Paul Van Delst (xmerge.f):

1) These conditions did not find the proper bounds all the time:
   Change: IF (VIDVEM.LT.VIDVRF.AND.VIDVEM.LE.VIDVBD) IEMBB = 1
           IF (VIDVRF.LT.VIDVEM.AND.VIDVRF.LE.VIDVBD) IEMBB = 2
   To: IF (VIDVEM.LE.VIDVRF.AND.VIDVEM.LE.VIDVBD) IEMBB = 1
       IF (VIDVRF.LE.VIDVEM.AND.VIDVRF.LE.VIDVBD) IEMBB = 2

2) Change this test (does not impact this error).
   Change: IF ((NELMNT.LE.0).OR.(NELMNT.GE.NLIMEM)) THEN
   To:     IF ((NELMNT.LT.0).OR.(NELMNT.GE.NLIMEM-1)) THEN

3) When reading from the emissivity file the array index was off by one:
   Change: CALL LINTCO(V1A,ZEMIS(NELMNT),V1B,ZEMIS(NELMNT+1),VI,ZINT,
   To:  CALL LINTCO(V1A,ZEMIS(NELMNT+1),V1B,ZEMIS(NELMNT+2),VI,ZINT,

4) There is also another bug, but it does not look like it was being
   used for anything.
   Change:   V=VI+FLOAT(J-1)*DVPO
   To:       V=V1PO+FLOAT(J-1)*DVPO
   (in two places)

Then repeated the changes for the reflectivity.

MWS 05/02 and SAC 07/02

Added Lambertian surface reflection option into LBLRTM making the diffusivity approximation
Is an approximation to the downwelling flux.  It uses the Lambertian or 'diffusivity' approximation
assuming  isotropic scattering with a value obtained from the diffusivity angle
corresponding to diffuse_fac = 1.67 (secant)
(xmerge.f)

MWS and SAC 07/02

Added in modifications to chi factor setting it equal to 1/2 its value
and modified the CO2 continuum by increasing it by a factor
of 7 (oprop.f and contnm.f).

MWS 07/02

Fixed small F4FN line coupling problem pointed out by
Frank Evans: SPP(I) = -SPP(I) (oprop.f)

SAC 07/02

Rayleigh scattering was included twice (lbllow.f and contnm.f),
pointed out by Minzheng Duan:

In subroutine TRANS, the SCT has be define as (line 2430 in LBLLOW.F)
         SCT = SCT+SCTMOL
But when calculating weighted asymmetry factor, the following
formula is used:(line 2462 in LBLLOW.F)
         ASYDM(NV) = ASYDM(NV)+SCT+SCTMOL

Tony's  fix:
c     The effects of Rayleigh scattering are also included in module
c     lbllow.f with the aerosol/cloud properties.  In the case that
c     that lbllow.f is selected (iaersl .ne. 0), the decision has been
c     made to include this effect with the scattering processes,
c     even though it is molecular in nature.  Otherwise the effects
c     of Rayleigh scattering are included in contnm.f.

MWS 07/02

To make compatible with PGI LINUX compiler (both f77 and f90)
we :
(1) removed the variable and cancatenation out of the stop command
(2) made the option of declaring IOSTAT in OPEN to be explicitly integer*4 iostat for
    the compiler:  (lblrtm.f, oprop.f, oprop_voigt.f)

MWS 08/02

Globally changed in many of the subroutines:
IFIX to INT
FLOAT to REAL
SNGL to REAL
AMAX1 to MAX
AMIN1 to MIN

Note: changes to oprop were also transcripted to
      oprop_voigt, nonlte, and nonlte_voigt but not tested