LBLRTM_v10.5
****************** lblatm       ******************

Number of input levels (sonde levels) has been increased to 6000

****************** lbllow       ******************

1) Number of input levels (sonde levels) has been increased to 6000.

2) ISEED has been set Integer*4 to retain consistency with subroutine.

3) A test has been added for ikmax to provide a proper result for a
   horizontal (or single) layer with aerosols

****************** lblrtm       ******************

1)  Number of input levels (sonde levels) has been increased to 6000.

2)  Capability to scale cross section (xs) profiles has been added.

3)  Incorect implementation of IMRG 40 and 41 (42 and 43) option  for
    Analytic Jacobians has been corrected.

4)  A proper implementation for the upwelling radiance has been implemented,
    IMRG =41 (43).

5)  The common block pathd has been made consistent by adding
<       character*4 ht1,ht2
    at line 4020

****************** nonlte       ******************

The identification parameters have been made consistent with the lblrtm naming convention.

****************** nonlte_voigt       ******************

The identification parameters have been made consistent with the lblrtm naming convention.

****************** oprop        ******************

Number of input levels (sonde levels) has been increased to 6000.

****************** oprop_voigt ******************

Number of input levels (sonde levels) has been increased to 6000.

****************** pltlbl       ******************

A change was made to enable execution with g95 (2 Aug 2006 version)

****************** postsub      ******************

A change was made to enable the proper opening/closing of file ifile

****************** solar        ******************

Common Block /BUFPNL/ was changed to /BUFPNL_s/:
/BUFPNL_s/ is used in the solar module and has a different length than /BUFPNL/.

Variable specifications were modified to be consistent within the module
and to ensure proper performance with the single precisio solar file
for both single and double versions of lblrtm.

****************** xmerge       ******************

Number of input levels (sonde levels) has been increased to 6000.

A change was made in an if block to facilitate comprehension.

EXTID was made character to ensure consistency of the variavble with its use.

Added:
<       dimension xdn_dum(2),tr_dum(2)

Coding modified to ensure proper calculation with Imrg 41
including computation of the reflected radiation at the surface.


**************************************************